Implement handlers composition
Now a request handler can be implement by chaining operators
This works as long as the method doesn't need to return anything.
For example `withMethod(method: Method)` just filters requests
and so doesn't need to return anything. `withRoute` instead returns
a `Args` so it can't implemented using this chaning operators, so
it can be used only to close a chain.
The implementation adds a boolean flag in the Request object to
allow to run `.next()` on a request and avoid running the code in
other chaining operators. To achieve this `next()` now will run
asynchronously on the event loop and it will set `nextRequested = true`
and unset it right after execution the next handler.