Move EventPollingExecutorScheduler to separate module
If it is in the same module it ends up being in the classpath and can
create linking errors depending on the order of dependencies (it happens
if you put `snunit-http4s` before `snunit-async-epollcat` in your
classpath.
The solution is to move the Java file in a separate module which is a
compile only dependency.